Break up large files into well-organized modules
Split 4 oversized files (4676 lines total) into 26 focused modules:
- Settings.tsx (1720 lines) → 9 files in components/settings/
- TerminalPane.tsx (1338 lines) → 8 files in components/terminal-pane/
- pane-manager.ts (1060 lines) → 6 files in lib/pane-manager/
- TabBar.tsx (558 lines) → 3 files in components/tab-bar/

All new files pass the max-lines oxlint rules (300 TS / 400 TSX).
No barrel exports, no util/helper naming.
